LPS-75955 Move MBMessageIndexer and MBThreadIndexer to message boards module
