+++ portal-impl/src/META-INF/message-boards-spring.xml
@@ -12,7 +12,5 @@
 	<bean class="com.liferay.portlet.messageboards.service.permission.MBMessagePermission" id="com.liferay.portlet.messageboards.service.permission.MBMessagePermission" />
 	<bean class="com.liferay.portlet.messageboards.service.permission.MBThreadPermission" id="com.liferay.portlet.messageboards.service.permission.MBThreadPermission" />
 	<bean class="com.liferay.portlet.messageboards.service.permission.MBPermission" id="com.liferay.portlet.messageboards.service.permission.MBPermission" />
-	<bean class="com.liferay.portlet.messageboards.util.MBMessageIndexer" id="com.liferay.portlet.messageboards.util.MBMessageIndexer" />
 	<bean class="com.liferay.portlet.messageboards.util.MBOpenSearchImpl" id="com.liferay.portlet.messageboards.util.MBOpenSearchImpl" />
-	<bean class="com.liferay.portlet.messageboards.util.MBThreadIndexer" id="com.liferay.portlet.messageboards.util.MBThreadIndexer" />
 </beans>
